Introduce IndexMetadataExternal. Refactoring, cleanups.
