+++ PLAN.md
@@ -24,7 +24,9 @@ Integrating with Solr will involve the following steps.
   - [ ] PROBLEM: we normally only store the first value indexed at a position in the forward index. But with the integrated index, we can't tell what value was first anymore by the time we write the forward index (because we reverse the reverse index). So we should either change it to store all values in the forward index (doable but complicated and doesn't tell us how to build concordances), or we need to add extra information to the tokens (payload?) that tells us what value was first. Probably refer to Mtas (they do store all values in the forward index, but they still need to know what token to use for concordances).
   - [ ] ForwardIndexAccessorLeafReader: implement per-segment terms class (or as part of ForwardIndexSegmentReader). Use same approach as global terms when comparing insensitively.
   - [ ] Only create forward index for annotations that should have one (fieldsconsumer)
+        (need access to index metadata)
   - [ ] Improve how we decide what Lucene field holds the forward index for an annotation (which sensitivity)
+    (need access to index metadata)
   - [ ] Speed up index startup for integrated index. Currently slow because it determines global term ids and sort positions dynamically. Unfortunately, there is no place in the Lucene index for global information, so this is a challenge. Maybe we can store the sort positions per segment and determine global sort positions by merging the per-segment lists efficiently (because we only need to compare strings if we don't know the correct order from one of the segment sort positions lists)<br/>
    (Ideally, we wouldn't need the global term ids and sort positions at all, but would do everything per segment and merge the per-segment results using term strings.)
 - [ ] Make indexmetada.yaml part of the Lucene index. Probably take the opportunity to refactor and simplify related code as much as possible. E.g. use Jackson, get rid of old settings, don't try to autodetect stuff from the index, etc.
