+++ engine/src/main/java/nl/inl/blacklab/search/indexmetadata/Annotation.java
@@ -13,11 +13,6 @@ public interface Annotation {
 
 	/** @return this annotation's name */
 	String name();
-	
-	/** Get the subannotation name, if this is a subannotation.
-	 * @return subannotation name, or null if this is not a subannotation
-	 */
-    String subName();
     
     /**
      * Is this a subannotation?
@@ -99,25 +94,9 @@ public interface Annotation {
     default String forwardIndexIdField() {
         return AnnotatedFieldNameUtil.forwardIndexIdField(luceneFieldPrefix());
     }
-    
-    /**
-     * Get subannotation descriptor.
-     * 
-     * Only valid for old-style indexes! (IndexMetadata.subannotationsStoredWithParent() == true)
-     * 
-     * Note that subannotations are not (yet) declared in index structure,
-     * so this will always succeed, even if the subannotation wasn't actually
-     * indexed. In that case, no hits will be found.
-     * 
-     * @param subName subannotation name
-     * @return subannotation descriptor
-     */
-    Annotation subannotation(String subName);
 
     /**
      * Get names of the subannotations for this annotation.
-     * 
-     * Only valid for new-style indexes! (IndexMetadata.subannotationsStoredWithParent() == false)
      * 
      * @return names of annotations that are considered subannotations of this annotation
      */
@@ -128,21 +107,6 @@ public interface Annotation {
     
     @Override
     int hashCode();
-
-    /**
-     * Return prefix for the value we're searching for, if any.
-     * 
-     * We used to index subannotations in the same field as their parent annotation,
-     * with the values prefixed. We don't do this anymore, but for old indexes,
-     * this is still relevant.
-     * 
-     * @return the prefix
-     */
-    default String subpropValuePrefix() {
-        if (isSubannotation() && indexMetadata().subannotationsStoredWithParent())
-            return AnnotatedFieldNameUtil.SUBANNOTATION_SEPARATOR + subName() + AnnotatedFieldNameUtil.SUBANNOTATION_SEPARATOR;
-        return "";
-    }
 
     void setSubAnnotation(Annotation parentAnnotation);
