+++ mocks/src/main/java/nl/inl/blacklab/mocks/MockAnnotation.java
@@ -92,16 +92,6 @@ public class MockAnnotation implements Annotation {
     }
 
     @Override
-    public Annotation subannotation(String subName) {
-        return null;
-    }
-
-    @Override
-    public String subName() {
-        return null;
-    }
-
-    @Override
     public boolean isSubannotation() {
         return false;
     }
