+++ server/src/main/java/nl/inl/blacklab/server/requesthandlers/RequestHandler.java
@@ -618,40 +618,41 @@ public abstract class RequestHandler {
 
     protected static void dataStreamMetadataGroupInfo(DataStream ds, BlackLabIndex index) {
         MetadataFieldGroups metaGroups = index.metadata().metadataFields().groups();
-        // FIXME: This synchronization is necessary because opening an index is not an atomic
-        //   operation, so it is apparently possible to get the metadata field groups object
-        //   before it's complete. We should fix the underlying problem here and make sure
-        //   index metadata is immutable.
-        synchronized (metaGroups) { // concurrent requests
-            Set<MetadataField> metadataFieldsNotInGroups = index.metadata().metadataFields().stream().collect(Collectors.toSet());
-            for (MetadataFieldGroup metaGroup : metaGroups) {
-                for (MetadataField field: metaGroup) {
-                    metadataFieldsNotInGroups.remove(field);
-                }
+        // TODO: This synchronization was necessary when testing many simultaneous
+        //   requests with Artillery. It should no longer be possible for metadata field
+        //   groups to change after the IndexMetadata object has been constructed though,
+        //   so it's been disabled. Verify this is using Artillery.
+        //synchronized (metaGroups) { // concurrent requests
+        Set<MetadataField> metadataFieldsNotInGroups = index.metadata().metadataFields().stream()
+                .collect(Collectors.toSet());
+        for (MetadataFieldGroup metaGroup : metaGroups) {
+            for (MetadataField field: metaGroup) {
+                metadataFieldsNotInGroups.remove(field);
             }
+        }
 
-            ds.startEntry("metadataFieldGroups").startList();
-            boolean addedRemaining = false;
-            for (MetadataFieldGroup metaGroup : metaGroups) {
-                ds.startItem("metadataFieldGroup").startMap();
-                ds.entry("name", metaGroup.name());
-                ds.startEntry("fields").startList();
-                for (MetadataField field: metaGroup) {
+        ds.startEntry("metadataFieldGroups").startList();
+        boolean addedRemaining = false;
+        for (MetadataFieldGroup metaGroup : metaGroups) {
+            ds.startItem("metadataFieldGroup").startMap();
+            ds.entry("name", metaGroup.name());
+            ds.startEntry("fields").startList();
+            for (MetadataField field: metaGroup) {
+                ds.item("field", field.name());
+            }
+            if (!addedRemaining && metaGroup.addRemainingFields()) {
+                addedRemaining = true;
+                List<MetadataField> rest = new ArrayList<>(metadataFieldsNotInGroups);
+                rest.sort(Comparator.comparing(a -> a.name().toLowerCase()));
+                for (MetadataField field: rest) {
                     ds.item("field", field.name());
                 }
-                if (!addedRemaining && metaGroup.addRemainingFields()) {
-                    addedRemaining = true;
-                    List<MetadataField> rest = new ArrayList<>(metadataFieldsNotInGroups);
-                    rest.sort(Comparator.comparing(a -> a.name().toLowerCase()));
-                    for (MetadataField field: rest) {
-                        ds.item("field", field.name());
-                    }
-                }
-                ds.endList().endEntry();
-                ds.endMap().endItem();
             }
             ds.endList().endEntry();
+            ds.endMap().endItem();
         }
+        ds.endList().endEntry();
+        //}
     }
 
     /**
