+++ text-pattern/src/main/java/nl/inl/blacklab/search/textpattern/TextPatternWildcard.java
@@ -28,7 +28,7 @@ public class TextPatternWildcard extends TextPatternTerm {
             return result.translate(context);
         try {
             return new BLSpanMultiTermQueryWrapper<>(QueryInfo.create(context.index(), context.field()), new WildcardQuery(new Term(context.luceneField(),
-                    context.subannotPrefix() + context.optDesensitize(optInsensitive(context, value)))));
+                    context.optDesensitize(optInsensitive(context, value)))));
         } catch (StackOverflowError e) {
             // If we pass in a really large wildcard expression,
             // stack overflow might occurs inside Lucene's automaton building
